refactor: Move 'core' elements to root of elements
