+++ src/syntax/elements/boxes/elementBoxIdentifier.ts
@@ -1,7 +1,7 @@
 import { TData, TDataName } from '@/@types/data';
 import { TElementName } from '@/@types/specification';
 
-import { ElementData } from '../core/elementArgument';
+import { ElementData } from '../elementArgument';
 import { getGlobalVariable } from '@/execution/symbolTable';
 
 // -------------------------------------------------------------------------------------------------
