+++ src/syntax/elements/elementSyntax.ts
@@ -1,6 +1,6 @@
 import { TDataName } from '@/@types/data';
 import { TElementName } from '@/@types/specification';
-import { IElementSyntax } from '@/@types/elementsCore';
+import { IElementSyntax } from '@/@types/elements';
 
 // -------------------------------------------------------------------------------------------------
