+++ src/syntax/specification.spec.ts
@@ -7,8 +7,8 @@ import {
 import specification from './specification';
 
 import { TData } from '@/@types/data';
-import { ElementData, ElementExpression } from './elements/core/elementArgument';
-import { ElementStatement, ElementBlock } from './elements/core/elementInstruction';
+import { ElementData, ElementExpression } from './elements/elementArgument';
+import { ElementStatement, ElementBlock } from './elements/elementInstruction';
 
 // -------------------------------------------------------------------------------------------------
